新版可查内存中的轨迹的代码

+++ src/main/java/com/ljj/flinkquery/demos/entity/newFive/firstResult.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 public class firstResult {
     int upCrowdNum;
     int downCrowdNum;
-    int upCrowdLength;
-    int downCrowdLength;
+    float upCrowdLength;
+    float downCrowdLength;
 }
